mockkubeapiserver: Refactor storage to be pluggable
This helps if we want e.g. to capture all the objects easily.
