+++ mockkubeapiserver/storage/resourceinfo.go
@@ -1,4 +1,4 @@
-package mockkubeapiserver
+package storage
 
 import (
 	"context"
@@ -7,44 +7,53 @@ import (
 
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
+	"k8s.io/apimachinery/pkg/runtime/schema"
+	"k8s.io/apimachinery/pkg/types"
+	"sigs.k8s.io/kubebuilder-declarative-pattern/mockkubeapiserver/forked"
 	"sigs.k8s.io/structured-merge-diff/v4/fieldpath"
 	"sigs.k8s.io/structured-merge-diff/v4/merge"
 	"sigs.k8s.io/structured-merge-diff/v4/typed"
-
-	"sigs.k8s.io/kubebuilder-declarative-pattern/mockkubeapiserver/forked"
 )
 
-type mockSchema struct {
-	builtin   *Schema
-	resources []*ResourceInfo
+// ResourceInfo exposes the storage for a particular resource (group-kind),
+// supporting CRUD operations for accessing the objects of that kind.
+type ResourceInfo interface {
+	GVK() schema.GroupVersionKind
+	ListGVK() schema.GroupVersionKind
+	ParseableType() *typed.ParseableType
+
+	GetObject(ctx context.Context, id types.NamespacedName) (*unstructured.Unstructured, bool, error)
+
+	ListObjects(ctx context.Context, filter ListFilter) (*unstructured.UnstructuredList, error)
+	Watch(ctx context.Context, opt WatchOptions, callback WatchCallback) error
+
+	CreateObject(ctx context.Context, id types.NamespacedName, u *unstructured.Unstructured) error
+	UpdateObject(ctx context.Context, id types.NamespacedName, u *unstructured.Unstructured) error
+	DeleteObject(ctx context.Context, id types.NamespacedName) (*unstructured.Unstructured, error)
 }
 
-func (s *mockSchema) Init() error {
-	schema, err := KubernetesBuiltInSchema()
-	if err != nil {
-		return err
-	}
-	s.builtin = schema
-	return nil
+type ListFilter struct {
+	Namespace string
 }
 
-type typeInfo struct {
-	ParserType typed.ParseableType
+type WatchOptions struct {
+	Namespace string
 }
 
-func (r *ResourceInfo) DoServerSideApply(ctx context.Context, live *unstructured.Unstructured, patchYAML []byte, options metav1.PatchOptions) (*unstructured.Unstructured, bool, error) {
-	if r.TypeInfo == nil {
-		return nil, false, fmt.Errorf("no type info for %v", r.GVK)
+func DoServerSideApply(ctx context.Context, r ResourceInfo, live *unstructured.Unstructured, patchYAML []byte, options metav1.PatchOptions) (*unstructured.Unstructured, bool, error) {
+	parserType := r.ParseableType()
+	if parserType == nil || !parserType.IsValid() {
+		return nil, false, fmt.Errorf("no type info for %v", r.GVK())
 	}
 
 	updater := merge.Updater{}
 
-	liveObject, err := r.TypeInfo.ParserType.FromUnstructured(live.Object)
+	liveObject, err := parserType.FromUnstructured(live.Object)
 	if err != nil {
 		return nil, false, fmt.Errorf("error parsing live object: %w", err)
 	}
 
-	configObject, err := r.TypeInfo.ParserType.FromYAML(typed.YAMLObject(patchYAML))
+	configObject, err := parserType.FromYAML(typed.YAMLObject(patchYAML))
 	if err != nil {
 		return nil, false, fmt.Errorf("error parsing patch object: %w", err)
 	}
@@ -63,7 +72,7 @@ func (r *ResourceInfo) DoServerSideApply(ctx context.Context, live *unstructured
 		return nil, false, fmt.Errorf("error encoding manager: %w", err)
 	}
 
-	apiVersion := fieldpath.APIVersion(r.GVK.GroupVersion().String())
+	apiVersion := fieldpath.APIVersion(r.GVK().GroupVersion().String())
 	mergedObject, newManagers, err := updater.Apply(liveObject, configObject, apiVersion, managers, string(managerJSON), force)
 	if err != nil {
 		return nil, false, fmt.Errorf("error applying patch: %w", err)
